requirements.txt
flask
google-genai
pillow
Send2Trash
